+++ framework-utilities/utilities/utilities-core/src/main/java/uk/gov/justice/services/common/configuration/JndiNameNotFoundException.java
@@ -1,0 +1,8 @@
+package uk.gov.justice.services.common.configuration;
+
+public class JndiNameNotFoundException extends RuntimeException {
+
+    public JndiNameNotFoundException(final String message) {
+        super(message);
+    }
+}