+++ framework-utilities/utilities/utilities-core/src/test/java/uk/gov/justice/services/common/util/StringProviderForTesting.java
@@ -1,0 +1,6 @@
+package uk.gov.justice.services.common.util;
+
+public interface StringProviderForTesting {
+
+    public String getString();
+}